+++ CHANGELOG.md
@@ -8,6 +8,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased] - ReleaseDate
 
+### Added
+
+* Added `Device::from_vid_pid`.
+
 ## [0.0.1] - ReleaseDate
 
 ### Added